+++ chatlogger.lua
@@ -8,13 +8,19 @@
 --]]
 
 function OnStartup()
-	tConfig, tChatHistory = {
+	tConfig, tChatHistory, tTopics = {
 		sBotName = SetMan.GetString( 21 ) or "PtokaX",
 		sProfiles = "012",		-- No history for commands from users with profiles
 		sLogsPath = "/www/ChatLogs/",
 		sTimeFormat = "[%I:%M:%S %p] ",
+		sPath = Core.GetPtokaXPath().."scripts/texts/",
+		sTickersList = "tickers.txt",
 		iMaxLines = 100,
-	}, { "Hi!" }
+		iTickerDelay = 6 * 60 * 60 * 10^3,		-- 6 hours to milliseconds
+		iTickerID = false,
+		iTopicIndex = 0,
+	}, { "Hi!" }, {}
+	ExecuteCommand.reloadtickers()
 end
 
 function ChatArrival( tUser, sMessage )
@@ -53,6 +59,17 @@ function UserConnected( tUser )
 	Core.SendToUser( tUser, sLastLines )
 end
 
+function OnTimer( iTimerID )
+	if tConfig.iTickerID ~= iTimerID then
+		return false
+	end
+	tConfig.iTopicIndex = ( tConfig.iTopicIndex + 1 ) % #tTopics
+	local tCurrentTopic, sUpdated = tTopics[ tConfig.iTopicIndex ], ( "<%s> Hub topic was updated by [ %%s ] to %%s." ):format( tConfig.sBotName )
+	SetMan.SetString( 10, tCurrentTopic.sTopic )
+	Core.SendToAll( sUpdated:format(tCurrentTopic.sNick, tCurrentTopic.sTopic) )
+	return true
+end
+
 RegConnected, OpConnected = UserConnected, UserConnected
 
 function History( iNumLines )
@@ -74,6 +91,11 @@ function LogMessage( sLine )
 	fWrite:write( sChatLine.."\n" )
 	fWrite:flush()
 	fWrite:close()
+end
+
+function CheckPermission( iProfile )
+	if not ProfMan.GetProfilePermission( iProfile, 7 ) then return false end
+	return true
 end
 
 function Reply( tUser, sMessage, bIsPM )
@@ -107,14 +129,61 @@ ExecuteCommand = {
 	topic = ( function()
 		local sErased, sUpdated = ( "<%s> Hub topic was erased by [ %%s ]." ):format( tConfig.sBotName ), ( "<%s> Hub topic was updated by [ %%s ] to %%s." ):format( tConfig.sBotName )
 		return function( tUser, sData, bIsPM )
-			if not ProfMan.GetProfilePermission( tUser.iProfile, 7 ) then return false end
-			if sData:len() == 0 then
+			if not CheckPermission( tUser.iProfile ) then return false end
+			if sData:len() == 0 or sData:lower() == "off" then
 				Core.SendToAll( sErased:format(tUser.sNick) )
-				SetMan.SetString( 10, "" )
-				return true
+				return ExecuteCommand.ticker( tUser, sData, bIsPM )
+			end
+			if tConfig.iTickerID then
+				TmrMan.RemoveTimer( tConfig.iTickerID )
+				tConfig.iTickerID = false
 			end
 			SetMan.SetString( 10, sData )
 			Core.SendToAll( sUpdated:format(tUser.sNick, sData) )
+			return true
+		end
+	end )(),
+
+	ticker = ( function()
+		local sReply = ( "<%s> Topics ticker has been activated." ):format( tConfig.sBotName )
+		return function( tUser, sData, bIsPM )
+			if not CheckPermission( tUser.iProfile ) then return false end
+			if tConfig.iTickerID then return false end
+			tConfig.iTickerID = TmrMan.AddTimer( tConfig.iTickerDelay )
+			return Reply( tUser, sReply, bIsPM )
+		end
+	end )(),
+
+	tickeradd = ( function()
+		local sFilePath, sReply, sTemplate = tConfig.sPath..tConfig.sFileName, ( "<%s> The topic has been added to tickers list." ):format( tConfig.sBotName ), "%s %s\n"
+		return function( tUser, sData, bIsPM )
+			if not CheckPermission( tUser.iProfile ) then return false end
+			local fTickerHandle, sNick, sTopic = io.open( sFilePath, "a+" ), sData:match "%-u (%S+) (.+)"
+			if not sNick then
+				sNick, sTopic = tUser.sNick, sData
+			end
+			fTickerHandle:write( sTemplate:format(sNick, sTopic) )
+			fTickerHandle:flush()
+			fTickerHandle:close()
+			return Reply( tUser, sReply, bIsPM )
+		end
+	end )(),
+
+	reloadtickers = ( function()
+		local sFilePath, sReply = tConfig.sPath..tConfig.sFileName, ( "<%s> Tickers list reloaded." ):format( tConfig.sBotName )
+		return function( tUser, sData, bIsPM )
+			local fTickerHandle = io.open( sFilePath )
+			if not fTickerHandle then
+				return false
+			end
+			for sLine in fTickerHandle:lines "*l" do
+				local sNick, sTopic = sLine:match "^(%S+) (.+)$"
+				table.insert( tTopics, {sNick = sNick, sTopic = sTopic} )
+			end
+			fTickerHandle:close()
+			if tUser then
+				Reply( tUser, sReply, bIsPM )
+			end
 			return true
 		end
 	end )(),